% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{blend_colors}
\alias{blend_colors}
\title{Blend Multiple Color Vectors}
\usage{
blend_colors(color_list, weights = NULL)
}
\arguments{
\item{color_list}{A list of character vectors containing color values in HEX format.
Each vector must be of the same length or length 1.}

\item{weights}{A numeric vector specifying the weight for each color vector.
If \code{NULL}, equal weights are assigned. The weights are normalized to sum to 1.}
}
\value{
A character vector of blended colors in HEX format.
}
\description{
Blends multiple color vectors by computing a weighted average of their RGB values.
}
\details{
This function takes multiple color vectors and blends them based on a weighted average
of their RGB values. If a color vector has length 1, it is recycled to match the longest
vector. If \code{weight} is provided, it must be numeric and match the number of color vectors.
Otherwise, equal weights are used.
}
\examples{
blend_colors(
  list(c("red", "blue"), c("yellow", "green")),
  weights = c(0.7, 0.3)
)

blend_colors(
  list("red", "blue", "yellow"),
  weights = c(0.4, 0.4, 0.2)
)

}
