% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_agreement_metrics.R
\name{dev_from_med}
\alias{dev_from_med}
\title{Mean absolute deviation from the median}
\usage{
dev_from_med(ratings)
}
\arguments{
\item{ratings}{Numeric vector of ratings (length >= 1)}
}
\value{
Numeric mean absolute deviation
}
\description{
Mean absolute deviation from the median
}
