% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_agreement_metrics.R
\name{modal_prop}
\alias{modal_prop}
\title{Proportion of ratings in the modal category}
\usage{
modal_prop(ratings)
}
\arguments{
\item{ratings}{Numeric vector of ratings (length >= 1)}
}
\value{
Numeric proportion in \link{0,1}
}
\description{
Proportion of ratings in the modal category
}
