% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleMriControl.R
\name{moduleMriControlServer}
\alias{moduleMriControlServer}
\title{\if{html}{\out{<i class="fas fa-circle-notch" style="font-size: 1.5em;" data-toggle="tooltip" title="Outline">}}\if{html}{\out{</i>}}' = "outline",}
\usage{
moduleMriControlServer(
  id,
  mri_sag_out,
  mri_axi_out,
  mri_cor_out,
  voxel_df_input,
  mode_init = "selection",
  external_selection = function() {
     NULL
 },
  external_selection_opts = list()
)
}
\description{
\if{html}{\out{<i class="fas fa-circle-notch" style="font-size: 1.5em;" data-toggle="tooltip" title="Outline">}}\if{html}{\out{</i>}}' = "outline",
}
