% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleMriPlane.R
\name{moduleMriPlaneUI}
\alias{moduleMriPlaneUI}
\title{\if{html}{\out{<i class="fas fa-ruler-horizontal" style="font-size: 1.5em;" data-toggle="tooltip" title="Scale Bar">}}\if{html}{\out{</i>}}' = "scale_bar" # not necessary?}
\usage{
moduleMriPlaneUI(
  id,
  plane,
  title = "MRI",
  slider_min = 1,
  slider_max = 256,
  pd = 5,
  siamp = 30,
  showValue = TRUE,
  wmp = 500,
  cf = F
)
}
\description{
\if{html}{\out{<i class="fas fa-ruler-horizontal" style="font-size: 1.5em;" data-toggle="tooltip" title="Scale Bar">}}\if{html}{\out{</i>}}' = "scale_bar" # not necessary?
}
