% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_agreement_metrics.R
\name{pair_exact_agreement}
\alias{pair_exact_agreement}
\title{Proportion of pairwise exact agreement}
\usage{
pair_exact_agreement(ratings)
}
\arguments{
\item{ratings}{Numeric vector of ratings (length >= 2)}
}
\value{
Numeric proportion in \link{0,1}
}
\description{
Proportion of pairwise exact agreement
}
