% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_file.R
\docType{class}
\name{ProgressCB-class}
\alias{ProgressCB-class}
\alias{ProgressCB}
\title{ProgressCB Class}
\description{
S4 class to store and track score assignment progress.
}
\section{Slots}{

\describe{
\item{\code{link}}{The weblink under which to store the progress.}

\item{\code{score_assigned}}{A data.frame indicating the scores assigned by the user.
Required columns: CBscore, x, y, z. (x,y,z are used for left_join())!}

\item{\code{user_affiliation}}{A character string specifying the user's institutional affiliation.}

\item{\code{user_name}}{A character string with the name of the user.}
}}

