% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns_agreement_metrics.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Normalized Shannon entropy consensus index}
\usage{
shannon_entropy(ratings)
}
\arguments{
\item{ratings}{Numeric vector of ratings (length >= 1)}
}
\value{
Numeric consensus index in \link{0,1}
}
\description{
Normalized Shannon entropy consensus index
}
